+++ base/dates/ranges.jl
@@ -28,7 +28,7 @@ Base.steprem(a::T, b::T, c) where {T<:TimeType} = b - (a + c * len(a, b, c))
 import Base.in
 function in(x::T, r::StepRange{T}) where T<:TimeType
     n = len(first(r), x, step(r)) + 1
-    n >= 1 && n <= length(r) && r[n] == x
+    n >= 1 && n <= length(r) && isequal(r[n], x)
 end
 
 Base.start(r::StepRange{<:TimeType}) = 0