+++ NEWS.md
@@ -141,6 +141,8 @@ Breaking changes
 
 This section lists changes that do not have deprecation warnings.
 
+  * `in` now uses `isequal` to test for containment ([#9381]).
+
   * `getindex(s::String, r::UnitRange{Int})` now throws `UnicodeError` if `last(r)`
     is not a valid index into `s` ([#22572]).